+++ CHANGELOG.md
@@ -5,7 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - Add loading screen messages
-- Show Hero Power details on mouse over
+- Show Hero Power and Weapon details on mouse over
 - Various loading screen strings
 
 ### Fixed